homework2_4
In this step, the sign up form was defined, as well as the personalized
page of the user, containing the default favorite link, together with
the image that has attached a  link to the weather website. As well as
this default first page after logging in, we have defined a default
account for the user with the name dev.